Rename Keep_Response_File to Keep_Temporary_Files.
And use this new variable to control whether the list of symbols to
export from a shared library is deleted or not.

Motivated by OB26-025.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@243399 936e1b1b-40f2-da11-902a-00137254ae57

+++ src/gprlib-build_shared_lib.adb
@@ -399,26 +399,28 @@ procedure Build_Shared_Lib is
 
       --  Delete temporary files
 
-      declare
-         Dont_Care : Boolean;
-         pragma Warnings (Off, Dont_Care);
-      begin
-         --  Delete response file, if any, except when asked not
+      if Delete_Temporary_Files then
+         declare
+            Dont_Care : Boolean;
+            pragma Warnings (Off, Dont_Care);
+         begin
+            --  Delete response file, if any, except when asked not
 
-         if Response_File_Name /= No_Path and then Delete_Response_File then
-            Delete_File (Get_Name_String (Response_File_Name), Dont_Care);
+            if Response_File_Name /= No_Path then
+               Delete_File (Get_Name_String (Response_File_Name), Dont_Care);
 
-            if Response_2 /= No_Path then
-               Delete_File (Get_Name_String (Response_2), Dont_Care);
+               if Response_2 /= No_Path then
+                  Delete_File (Get_Name_String (Response_2), Dont_Care);
+               end if;
             end if;
-         end if;
 
-         --  Delete export file
+            --  Delete export file
 
-         if Standalone /= No and then Export_File /= No_Path then
-            Delete_File (Get_Name_String (Export_File), Dont_Care);
-         end if;
-      end;
+            if Standalone /= No and then Export_File /= No_Path then
+               Delete_File (Get_Name_String (Export_File), Dont_Care);
+            end if;
+         end;
+      end if;
 
       if not Success then
          if Driver_Name = No_Name then
